+++ src/placo/planning/walk_pattern_generator.cpp
@@ -195,7 +195,7 @@ void WalkPatternGenerator::planCoM(Trajectory& trajectory, Eigen::Vector2d initi
   // Computing how many steps are required
   int timesteps = 0;
 
-  for (int i = 0; i < trajectory.supports.size(); i++)
+  for (size_t i = 0; i < trajectory.supports.size(); i++)
   {
     timesteps += support_dt(trajectory.supports[i]);
 
@@ -226,7 +226,7 @@ void WalkPatternGenerator::planCoM(Trajectory& trajectory, Eigen::Vector2d initi
   // Adding ZMP constraint and reference trajectory
   int constrained_timesteps = 0;
   FootstepsPlanner::Support current_support;
-  for (int i = 0; i < trajectory.supports.size(); i++)
+  for (size_t i = 0; i < trajectory.supports.size(); i++)
   {
     current_support = trajectory.supports[i];
     int step_timesteps = support_dt(current_support);
@@ -298,7 +298,7 @@ void WalkPatternGenerator::planFeetTrajectories(Trajectory& trajectory, Trajecto
     trajectory.yaw(side).addPoint(t, old_trajectory->yaw(side).get(t), 0, true);
   }
 
-  for (int step = 0; step < trajectory.supports.size(); step++)
+  for (size_t step = 0; step < trajectory.supports.size(); step++)
   {
     auto& support = trajectory.supports[step];
